Update sheodox-ui, fix 500 error creating new activity lists
The sheodox-ui update was required to make RichText parse things more safely, otherwise malformed links could crash the page.

Also made sure we're passing an array when on the new activity list form, otherwise you can't create a list as it tries to do array operations on an empty array.